fix: issue #42 second comment: export A and clear causes memory problem because of val is null
