+++ src/main.c
@@ -71,6 +71,7 @@ int	main(int argc, char **argv, char **envp)
 		}
 	}
 	free_dict(env_dict);
-	rl_clear_history();
+//	rl_clear_history();
+	clear_history();
 	return (0);
 }
